mcp_server/src/requirements.txt
mcp[cli]
requests 
pyjwt
cryptography
